save_plant: inelegant concurrency fix
This inelegant and hackyish fix gives a randomly generated name to the
temporary savefiles, so there is only a 0.1% chance of the filename
being the same twice in a row.

This solves the ocasional exceptions raised in save_plant, when there
is more than one call still running at once, for the same user (due to
slow writes).